+++ source/app/service-providers/commands/export.ts
@@ -19,7 +19,7 @@ import { trans } from '@common/i18n-main'
 import type { ExporterOptions } from './exporter/types'
 import { promises as fs } from 'fs'
 import path from 'path'
-import { PANDOC_WRITERS } from '@common/util/pandoc-maps'
+import { PANDOC_WRITERS } from '@common/pandoc-util/pandoc-maps'
 import { type PandocProfileMetadata } from '@providers/assets'
 import { runShellCommand } from './exporter/run-shell-command'
 import { showNativeNotification } from '@common/util/show-notification'
