+++ source/app/service-providers/commands/importer/index.ts
@@ -25,7 +25,7 @@ import { app, dialog } from 'electron'
 import { trans } from '@common/i18n-main'
 import type AssetsProvider from '@providers/assets'
 import { type PandocProfileMetadata } from '@providers/assets'
-import { SUPPORTED_READERS } from '@common/util/pandoc-maps'
+import { SUPPORTED_READERS } from '@common/pandoc-util/pandoc-maps'
 import { hasMarkdownExt } from '@common/util/file-extention-checks'
 
 export default async function makeImport (
