+++ source/win-assets/DefaultsTab.vue
@@ -95,7 +95,7 @@ import ZtrAdmonition from '@common/vue/ZtrAdmonition.vue'
 import { trans } from '@common/i18n-renderer'
 import { ref, computed, toRef, watch, onUnmounted } from 'vue'
 import type { AssetsProviderIPCAPI, PandocProfileMetadata } from '@providers/assets'
-import { PANDOC_READERS, PANDOC_WRITERS, SUPPORTED_READERS } from '@common/util/pandoc-maps'
+import { PANDOC_READERS, PANDOC_WRITERS, SUPPORTED_READERS } from '@common/pandoc-util/pandoc-maps'
 import sanitizeFilename from 'sanitize-filename'
 import { DateTime } from 'luxon'
 import { parseReaderWriter } from 'source/common/pandoc-util/parse-reader-writer'
