+++ source/win-project-properties/App.vue
@@ -164,7 +164,7 @@ import ZtrAdmonition from '@common/vue/ZtrAdmonition.vue'
 import { ref, computed, watch } from 'vue'
 import type { ProjectSettings, DirDescriptor, AnyDescriptor, MDFileDescriptor, CodeFileDescriptor } from '@dts/common/fsal'
 import type { AssetsProviderIPCAPI, PandocProfileMetadata } from '@providers/assets'
-import { PANDOC_READERS, PANDOC_WRITERS, SUPPORTED_READERS } from '@common/util/pandoc-maps'
+import { PANDOC_READERS, PANDOC_WRITERS, SUPPORTED_READERS } from '@common/pandoc-util/pandoc-maps'
 import { type WindowTab } from '@common/vue/window/WindowTabbar.vue'
 import { useConfigStore } from 'source/pinia'
 import objectToArray from 'source/common/util/object-to-array'
